deploy: use allowedTopologies in storage classes
This is redundant when the scheduler extensions are active, but when
those are disabled, then allowedTopologies prevents scheduling
of pods onto nodes where the driver is not running.

The central provisioner recovered from that by removing the bad
"selected node" annotation. With distributed provisioning such a pod
would get stuck forever.

Testing of the scenario without scheduler extensions is covered by
adding late binding tests to "operator-direct-production". Using an
alternative driver name is covered by "operator-lvm-production".