test: fix operator event checking
The code had two problems.

It stored pointers to the objects passed in by the caller, which is
dangerous because those instances might be reused by the caller, in
which case the content changes. In practice this doesn't seem to
happen, but we can't know that.

It relied on receiving events in a certain order. The apiserver
and (to some extend) the fake client may combine events, which
apparently can also cause them to be delivered out-of-order. This has
happened in practice and broke the
test (https://github.com/intel/pmem-csi/issues/852).

The solution is to de-duplicate after sorting. Because the order
depends on what we compare against, it is done during the actual
comparison together with de-duplication.