operator: set operator's log verbosity level
Operator running in e2e tests was not showing any logs, and the reason is
that there is no log verbosity level set for the operator deployment.
The default level is 0 which suppress all informative logs.

The generated operator manifest is set to log level 3, and the test
script updates it with the TEST_OPERATOR_LOGLEVEL environment value,
which defaults to 5.