operator: remove objectPatch
The objectPatch struct obscured the flow in redeploy more than it
helped because object manipulations were scattered through various
functions.

Also, to/from was easy to confuse. As a result, the parameters in
deployment_controller.go were backwards and the diff that it showed
was from new object to the old.