test/deployment: stop generating node secrets/certificates
Node controller secrets are no more used by the driver, so it is
unnecessary to generate them. The only secret/certificates the
controller driver required is the webhook secrets.

This shouldn't be a problem for running version skew tests as those
tests uses the version specific deployment script. So downgrading the
driver version to v0.8 requires those secrets and they gets generated by
the setup scripts of that revision of code.