test: allow webhook failures during testing
The webhook was kept as mandatory for testing when relaxing the
default, with the rationale being that we want to know if it starts
failing. But we do get test flakes because of that. Normally, pod
creation would get retried, it's just that many E2E tests don't do
that.

It's better to let the tests run and cover testing of the webhook
through WaitForPMEMDriver which explicitly creates a pod and verifies
that the webhook was invoked.