pmem-csi-driver: optionally convert raw namespaces
The new functionality gets added to the existing binary to minimize
overall image size.

Unit testing with a fake ndctl implementation is not very realistic
because it does not mimick all constraints of the actual library, but
it's still better than none and can be used at least to verify the
behavior when invoked in different hardware environments, something
that would be difficult in a E2E testing.