deploy: convert raw namespaces in DaemonSet
This creates the DaemonSet which runs on nodes with the "<driver
name>/convert-raw-namespaces: force" label. It's inactive if that
label is never set by a user.

Node pods are now allowed to run on the master node. This is usually
not necessary and won't be triggered because the required labels won't
be set for that node, but if someone wants that, we should allow it,
i.e. tolerate the node taint.

The immediate usage for this will be testing of raw namespace
conversion on the master node of the QEMU cluster.