pmem-csi-driver: create volume group as part of conversion
For legacy PMEM we cannot rely on the alt name. When conversion is
explicitly requested, we can and have to create the expected volume
group directly, then the normal driver will use it without checking
the alt name.

For the sake of consistency with the normal mode, only active regions
are considered. This made no difference so far, but if we ever really
encounter an inactive region, then we probably cannot use it anyway
because our code does not activate it.