pmem-csi-driver: user error if driver cannot run
When a user explicitly asks for conversion, the expected outcome is
that afterwards the PMEM-CSI driver can run in LVM mode. By checking
that explicitly and throwing an error if it is not the case instead
of silently not starting the LVM driver the user can investigate by
reading the logs of the failed conversion pod.