exec: cleanly separate stdout/stderr
Commit 40ba3555a43fa2aa8aa1f081af1f2386edda4dd8 intentionally captured
stderr in the output because earlier it was hard to understand why LVM
commands failed. However, we now learned the some error messages can
be safely ignored and not doing so broke LVM mode: when /dev/sda is
unreadable, "vgs" prints a line about it on stderr, but the output on
stdout is fine.

Now we capture both output streams separately, log both, but only
return stdout to the caller.

The TestResult call is unrelated to that fix and will be used for
additional debugging in another PR. It gets added here because it can
use the same test cases and is related.

(cherry picked from commit 3a7c35411a3931085ef001f81d6b4cce3358e01d)