operator: avoid unexpected object patching
On redeploying the objects operator should not unset the defaults chosen
by the API server, otherwise it results in patch diff and leads to
unnecessary update object call. This is observed both driver
statefulset and daemonset objects.

To avoid this we have two options:
 1. set only specific options instead of total refresh.
 2. set also the default values

We chose option 1 while setting daemonset and statefulset specification
fields, and option 2 for setting container values.

FIXES #924