test: operator version-skew tests using OLM
The `operator-sdk run bundle-upgrade` supports upgrading the running
operator to a new version. Currently, it supports only upgrading, though
it supposed to work same for downgrade but, we are not considering that
for now.

The test first generates the bundle-manifests for the downgraded
release(say v0.9) source and deploys the operator using olm. And then,
it regenerates the bundle manifests for devel source giving the current
version(say 1.0) number to the bundle and redeploys the operator with
canary image.