test/e2e: fix recover tests
We should not wait for 'start/stop' metrics up on sub-object update as
that does not lead to the deployment reconcile. And we can ignore the
object updates, as one sub-object change might lead to change in other
dependent object.

So we only validate the driver once after the object recover.