e2e: remove socat service and node port
Instead of connecting to the service via the node IP address, we can
also connect to the pod directly via our new pod dialer. The node port
was problematic because reinstalling the driver sometimes got stuck
when Kubernetes did not allow reuse of the port quickly enough.

One side effect is that this removes a reference to code in
csi-test/v3.