nodeserver: Improve idempotency in ephemeral volume creation
The check for existing file system before mkfs was made on
one call path only, but ephemeral took another path when creating
an inline volume. Due to 2 additional side effects: incomplete
idempotency testing, and issue in kubelet that caused repeated
publishVolume requests, it was exposed that nodeserver attempts
to mkfs a device with existing file system, which fails,
and there is no useful error information.
dax flag gets reported as "dax=always" which means we have
to use special handling for comparing.