docs: discourage usage of "devel" documentation
Users who landed on https://intel.github.io/pmem-csi/ or used
https://intel.github.io/pmem-csi/latest were told how to install the
development version of PMEM-CSI and then probably unintentionally
tried that instead of using the latest stable release.

We now make the latest stable release the main page, with "devel" in
the URL for the development version, "latest" as symlink for the
latest stable version (offered for those who still have those pages
bookmarked or in emails).

When pointing to documentation, we should always use the versioned
pages because those URLs are stable.