docs: revise installation, examples, troubleshooting
We want users to use late binding, therefore our example should also
use it.

When reading from top to bottom, storage classes should be explained
before using them.

Troubleshooting was partially covered under "install from source" ("Is
my driver running?"), but better should be in a section that is
clearly marked as such. Several other cases are now also documented.

The examples section is now a stand-alone section, too, and can be
followed without checking out the repository.

"Install from source" gets renamed to "Install via YAML files" because
using the source code is a necessary evil (certificates are needed),
but its not the main feature of that approach.