test: delete pod when it fails during startup
dax.CreateVolume may fail while checking the pod status. For the
caller this looks like the pod never got created and it thus
won't (and can't) delete it.

This problem showed up when Kata Containers failed to start the
pod. Because the pod wasn't deleted, the PVC was not deleted either
and the test timed out while trying to clean up.