pmem-csi-driver: fix NodePublishVolume for Kata Containers
NodePublishVolume expects volume parameters from CreateVolume in the
VolumeContext and used those to detect volumes meants for Kata
Containers. But CreateVolume never set up that context and thus
all volumes were treated as "normal" volumes.

NodeUnpublishVolume instead used the volume state file created by
CreateVolume and that did contain the flag for Kata Containers, with
the result that nodeUnpublishKataContainerImage failed:

   umount: /var/lib/pmem-csi.intel.com/mount/pvc-[...]: no mount point
   specified.

NodePublishVolume could use the state file, but the context is easier
and useful for logging, so CreateVolume gets fixed instead.