deploy: add liveness probes
Liveness probes serve as a fallback mechanisms for cases where the
container gets stuck but fails to terminate. We piggy-back onto the
metrics endpoint with HTTP probes because that endpoint is readily
available and, in the case of the PMEM-CSI node driver, already does
some work that mirrors what the driver needs to do during
operation (interact with underlying storage to retrieve capacity).

A startup probe is used because the node driver might start up more
slowly when setting up LVM. We can also add that for direct mode for
the sake of simplicity.

The liveness probes could be useful, although in practice the node
driver should already terminate if it runs into fatal issues, so we
probably won't need to rely on the probe. To avoid potential false
positives (killing the container although it only had a temporary
outage), the threshold is 5.

Readiness probes make less sense. The node driver does not do any load
balancing. It might make sense if we had more than one instance and
let the scheduler access only ready instances, but right now we only
have one instance.