deploy: test with "testing" deployments instead of "coverage"
The main motivation is that the "testing" deployments work on
OpenShift whereas the "coverage" deployments fail because they need a
hostPath volume, something that is not allowed (at least on one
OpenShift 4.6 cluster).

For upstream testing, we loose coverage of the pmem-csi-driver-test
image, but that is not a big loss because it's allowed to
break. Testing actually might run faster because we use the production
binaries.

We also loose coverage data collection, but that wasn't actually used
and can be brought back when needed.