docs: use modified linkcheck to detect stale links and anchors
The original linkcheck.py skips URLs that only consist of an anchor
but as we write those anchors manually, we do want to check them. It
doesn't know how to handle links to .md files and absolute links to
files or directories in the repo, leading to false negatives.

All of that gets added to a fork of the original linkcheck. This is
not something that can be upstreamed, unfortunately, because it is
based on assumptions about how PMEM-CSI documentation is generated (md
-> html, absolute links).

(cherry picked from commit 53d54c12e3b2fcbeb8f7db9e6d42472daac7b426)