docs: remove TOC from individual pages
Maintaining the TOC was a constant source of regressions with missing
or outdated entries. It's main purpose was for people browsing the
source code repo, but even there it might distract because one has to
jump over it to get to the actual text.

Users should read documentation on the docsite, and there we have the
sidebar.

(cherry picked from commit af24d70bc9380a9fe1926eb4e68cf7119fa14a8f)