pmem-csi-driver: improve volume size handling
It is better to handle zero-size volume requests inside the device
manager layer because the backend knows whether it supports those. The
caller shouldn't have to make assumptions about that.

Because the device manager may increase the size of a volume while
creating it, it must return the actual size for the volume.

With the new CSI spec we can and should return the maximum volume
size. This also makes it possible to write a more precise test that
covers storage fragmentation.

Printing int64 values in the GB range is often leading to very hard to
read numbers, in particular when trying to figure out whether the
decimal value is an exact multiple of GB. Using resource.Quantity will
automatically reduce to smaller numbers with a suitable Gi/Mi/Ki
suffix ("8Gi" instead of "8589934592").

Handling for ndctl_region_get_align return ULONG_MAX (= not know) was
missing.

The difference between actual and usable size of a namespace gets
clarified by providing two different functions for each.