pmem-csi-driver: fall back to 96MiB region alignment, refactor
Both ndctl and PMEM-CSI failed to create a 100MiB namespace with an
"invalid parameter" while trying to set the intended size on a QEMU VM
with
   -object memory-backend-file,id=memnvdimm0,mem-path=/nvme/gopath/src/github.com/intel/pmem-csi/_work/pmem-openshifttt-6bbxv-worker-0-2zzn8.nvdimm,prealloc=yes,size=1075838976,align=2097152
   -device nvdimm,node=0,label-size=2097152,memdev=memnvdimm0,id=nvdimm0,slot=0

The OS was CoreOS for OpenShift 4.8 alpha with Linux kernel
4.18.0-305.el8.x86_64.

It looks like that kernel fails to expose the region alignment. 96MiB
is used as default because we have seen it in some hardware
configurations (https://github.com/intel/pmem-csi/issues/964).

For unknown reasons the namespace creation code wasn't using
FsdaxAlignment(), in contrast to capacity retrieval. Now both use
the same helper function, so discrepancies like that should not occur
anymore.