pmem-csi-driver: support insecure scheduler endpoint
There are scenarios (for example, OpenShift) where configuring an
https URL in the scheduler simply cannot be done (lack of DNS,
certificate handling). In that case, a scheduler extender listening on
a node port without TLS is better than no scheduler extender.

The security risk associated with not using encryption is acceptable
because the information is not particularly sensitive (pod names and
node names) and what an attacker could learn from that is partly
available to anyone anyway (storage capacity on each node).