test: stop depending on host port for sanity testing
We can connect to the socat port via port forwarding. The advantage is
that the e2e.test can run on a host that has no IP routing to the
cluster nodes, it only needs access to the API server.

A secondary benefit is that we avoid potential conflicts because of
the host port.