test: fix broken operator-direct-production driver reuse
It used to be so that in this particular deployment the controller
wasn't installed because kube-system didn't have the necessary
secret. That changed when the controller also started to run the pod
rescheduler, with the result that the "expected" against "current"
deployment check always found a mismatch in HasController and always
restarted the driver for each test.