CI: fix suppression of 'make test_e2e' return code
Piping the output of the test command into other commands caused the
return code of the "make test_e2e" command to get lost, with the
result that the test was considered "unstable" (because of recorded
JUnit failures) instead of "failed" and later stages like publishing
images ran although they should have been skipped.

Also, the 2>&1 redirection probably was in the wrong place, with the
result that stderr was not filtered.

This is using a minimally intrusive solution where bash is selected
as interpreter by adding it at the start of the sh parameter, then
using pipefail.

This whole script probably should be rewritten with multiline strings,
but a naive attempt at that ended with several hours of trying
different syntax variants and still no success...