pmem-csi-driver: fix reuse of namespace in LVM mode
When the driver created namespaces for usage with LVM, it did not wipe
them directly after creation. If, for whatever reason, the new
namespace had LVM labels, all kinds of failures could occur.

Now "wipefs -a" is called to erase stale information. It is used
instead of overwriting blocks because the tool is readily available
and was written for that task.