test: LVM with multiple namespaces per VG
This ensures that the driver really creates new namespaces and adds
them to an existing VG. We had code for that in 0.9, but no tests, and
then temporarily disabled the feature.

The test accidentally also triggered an issue with not wiping the new
namespace when run multiple times. It's now intentionally run twice to
cover that scenario.