docs: replace manual text wrapping in table with automatic wrapping
We had plenty of tables with too much text to fit the screen. By
default, the ReadTheDocs theme makes such tables very wide with
scrollbars, which isn't very useful and might cause people to miss the
additional explanations.

We used manual line breaks in some places, but that's not a good
solution. For example, the 1.13 feature gates fit nicely into a single
line on my screen.

By overriding the theme settings we can get the browser to wrap text
for us. That's also how GitHub renders the tables.