operator: delete and re-create CSIDriver object
When updating a driver deployment on Kubernetes 1.21, the
CSIDriver.Spec.StorageCapacity field must be changed from false to
true. This cannot be done via patching because the object is
immutable. We have to fall back to deleting and re-creating it.

The v0.9.1 operator doesn't know that and fails during patching,
i.e. downgrades are not possible. This is not important, so we simply
skip this particular test scenario.