test: disable operator downgrade testing
It currently doesn't work even on Kubernetes 1.19. The scenario has
always been a bit questionable because there's no guarantee that the
old operator knows how to undo something created by the more recent
operator. Instead of investigating the exact reason, let's simply skip
the test.

For example, the switch to using a Deployment is not going to work for
downgrades because the old operator didn't support deleting them.