test: fix CA usage when deploying operator through OLM
Changing the directory in a script and not returning to the original
directory is a subtle change of global state and can lead to hard to
diagnose problems.

In this case, changing the current directory led to using a different
work directory when setting up secrets, with the result that the
operator used a new CA and the scheduler rejected connections.