pmem-csi: avoid false negative liveness probe
During a long-running DeleteVolume for a 200GB volume the LVM lock is held and
the liveness probe for metrics data timed out after a minute, killing the
container, because metrics retrieval must take the LVM lock.

As operation duration is non-deterministic, the solution is to probe only
metrics data which can be served immediately when the process is up and
running. This is less precise (= doesn't really check any functionality), but
that's an inherent problem of liveness probes: they simply cannot cover the
full functionality.