CI: improve log artifacts
Putting the Kubernetes version nearer the start of the file names ensures that
files for the same test case are grouped together when sorting. Sorting by the
type of content is less useful because usually one wants to investigate a
specific test failure which occurred under a specific Kubernetes version.

Splitting up log messages by node makes the individual files smaller and
simplifies debugging of a problem that occurred on a specific node.

To debug the spontanuous node
reboot (https://github.com/intel/pmem-csi/issues/1055) the full systemd journal
is needed.