e2e: fix OLM version skew testing
https://github.com/intel/pmem-csi/pull/1052 simplified the make rules
for the OLM bundle. However, that had the effect that those rules now
no longer can build the OLM bundle for the previous release, 1.0.1, which
caused version skew testing to fail.

The solution is to run make in the 1.0.1 source directory and thus use
the make rules from that release. This is cleaner, too.