build: collect coverage information in Jenkins
The support for this in the pmem-csi-driver-test still worked, we just have to
enable it before running tests. Running "make kustomize" with a different base
deployment achieves that. It modifies the current tree, but that is okay for
usage in Jenkins where the directory is going to be deleted.

Jenkins can consume Cobertura XML when the right set of plugins are
installed. How to do that was mostly based on trial-and-error, upstream
documentation was vague.