test: run OLM tests first, III
It seemed that Ginkgo v1 ran tests in the order in which they were defined, but
that only worked (once?) in a pull request, but not in branch testing. Perhaps
it was still random.

Ginkgo v2 has "Ordered" nodes that guarantee what we want: sequential execution
in the order of the Describe calls.