build: Kubernetes 1.24
The biggest change is that the corresponding klog now supports contextual
logging, so quite a bit of our custom code can be replaced.

https://github.com/kubernetes/kubernetes/commit/9e372bffeffbf52f024f23d050767bbeaa30ad92
made the result from NodeSSHHosts random, so we have to copy the relevant
code. This is ultimately the cleaner solution and will also enable future
enhancements, like supporting more than one master node.