test: Ginkgo v2.1.3
This is a breaking change for which we need updated Kubernetes and csi-test
releases. Ginkgo v2 is a major rewrite which simplifies some things (junit
report creation) and enables others (ordered execution of tests, reliable
shutdown).

Ginkgo gets invoked now so that we don't get escape sequences for
color (doesn't work well when viewing logs) and with less clutter (increased
slow spec threshold). A known issue is that -ginkgo.v causes output to be
printed twice (when it gets produced and when summarizing a spec, see
https://github.com/onsi/ginkgo/issues/967).